requirements.txt
requests>=2.28.0
Pillow>=9.0.0
# tkinter is built-in (conda install tk if missing)
